requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
requests==2.31.0
xmltodict==0.13.0
python-dotenv==1.0.0
gunicorn==21.2.0